refactor: move tj to environments folder
